[ObsUX][APM] Migration of Service Overview tests to deployment agnostic approach (#200226)
## Summary

Part of #193245
Closes #198986

This PR moves all compatible/supported test cases for Service Overview.
Unsupported cases are kept in the old test section to run on stateful
for now.

## How to Test

### Serverless

```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts --grep="APM API tests"
```

It's recommended to be run against
[MKI](https://github.com/crespocarlos/kibana/blob/main/x-pack/test_serverless/README.md#run-tests-on-mki)

### Stateful

```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts --grep="APM API tests"
```

---------

(cherry picked from commit 185fa2a7dc086a0b30925ea4447f9ec3de1c8651)
