+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/index.ts
@@ -38,5 +38,6 @@ export default function apmApiIntegrationTests({
     loadTestFile(require.resolve('./span_links'));
     loadTestFile(require.resolve('./suggestions'));
     loadTestFile(require.resolve('./throughput'));
+    loadTestFile(require.resolve('./service_overview'));
   });
 }
