+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/apm/service_overview/dependencies/es_utils.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { MappingTypeMapping } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
+import type { MappingTypeMapping } from '@elastic/elasticsearch/lib/api/typesWithBodyKey';
 import { v4 as uuidv4 } from 'uuid';
 
 export function createServiceDependencyDocs({
